Create a class to factorize HDF EOS file handler
