+++ satpy/readers/hdfeos_base.py
@@ -1,0 +1,100 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+# Copyright (c) 2019
+
+# This file is part of satpy.
+
+# satpy is free software: you can redistribute it and/or modify it under the
+# terms of the GNU General Public License as published by the Free Software
+# Foundation, either version 3 of the License, or (at your option) any later
+# version.
+
+# satpy is distributed in the hope that it will be useful, but WITHOUT ANY
+# WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+# A PARTICULAR PURPOSE.  See the GNU General Public License for more details.
+
+# You should have received a copy of the GNU General Public License along with
+# satpy.  If not, see <http://www.gnu.org/licenses/>.
+import logging
+
+from datetime import datetime
+
+from pyhdf.error import HDF4Error
+from pyhdf.SD import SD
+
+from satpy.readers.file_handlers import BaseFileHandler
+
+logger = logging.getLogger(__name__)
+
+
+class HDFEOSBaseFileReader(BaseFileHandler):
+    """Base file handler for HDF EOS data for both L1b and L2 products. """
+    def __init__(self, filename, filename_info, filetype_info):
+        super(HDFEOSBaseFileReader, self).__init__(filename, filename_info, filetype_info)
+        try:
+            self.sd = SD(self.filename)
+        except HDF4Error as err:
+            err_message = "Could not load data from file {}: {}".format(self.filename, err)
+            raise ValueError(err_message)
+        self.metadata = self.read_mda(self.sd.attributes()['CoreMetadata.0'])
+        self.metadata.update(self.read_mda(
+            self.sd.attributes()['StructMetadata.0'])
+        )
+        self.metadata.update(self.read_mda(
+            self.sd.attributes()['ArchiveMetadata.0'])
+        )
+
+    @staticmethod
+    def read_mda(attribute):
+        lines = attribute.split('\n')
+        mda = {}
+        current_dict = mda
+        path = []
+        prev_line = None
+        for line in lines:
+            if not line:
+                continue
+            if line == 'END':
+                break
+            if prev_line:
+                line = prev_line + line
+            key, val = line.split('=')
+            key = key.strip()
+            val = val.strip()
+            try:
+                val = eval(val)
+            except NameError:
+                pass
+            except SyntaxError:
+                prev_line = line
+                continue
+            prev_line = None
+            if key in ['GROUP', 'OBJECT']:
+                new_dict = {}
+                path.append(val)
+                current_dict[val] = new_dict
+                current_dict = new_dict
+            elif key in ['END_GROUP', 'END_OBJECT']:
+                if val != path[-1]:
+                    raise SyntaxError
+                path = path[:-1]
+                current_dict = mda
+                for item in path:
+                    current_dict = current_dict[item]
+            elif key in ['CLASS', 'NUM_VAL']:
+                pass
+            else:
+                current_dict[key] = val
+        return mda
+
+    @property
+    def start_time(self):
+        date = (self.metadata['INVENTORYMETADATA']['RANGEDATETIME']['RANGEBEGINNINGDATE']['VALUE'] + ' ' +
+                self.metadata['INVENTORYMETADATA']['RANGEDATETIME']['RANGEBEGINNINGTIME']['VALUE'])
+        return datetime.strptime(date, '%Y-%m-%d %H:%M:%S.%f')
+
+    @property
+    def end_time(self):
+        date = (self.metadata['INVENTORYMETADATA']['RANGEDATETIME']['RANGEENDINGDATE']['VALUE'] + ' ' +
+                self.metadata['INVENTORYMETADATA']['RANGEDATETIME']['RANGEENDINGTIME']['VALUE'])
+        return datetime.strptime(date, '%Y-%m-%d %H:%M:%S.%f')
